ref(replays): move start time in replay details (#67397)
Move the start time to be under the user info:

<img width="1268" alt="SCR-20240320-nwos"
src="https://github.com/getsentry/sentry/assets/56095982/4224b986-7c33-49af-bbf1-71cb5901aebe">

Based on Figma:
https://www.figma.com/file/bbTpXNRRu7NY4MDaAAFiJe/Specs%3A-Session-Replay-Mobile-Alpha-1?type=design&node-id=1-13024&mode=design&t=QKMYY86lUUQvBNcH-0
