+++ static/app/components/replays/header/replayMetaData.tsx
@@ -4,8 +4,7 @@ import styled from '@emotion/styled';
 
 import ErrorCounts from 'sentry/components/replays/header/errorCounts';
 import HeaderPlaceholder from 'sentry/components/replays/header/headerPlaceholder';
-import TimeSince from 'sentry/components/timeSince';
-import {IconCalendar, IconCursorArrow} from 'sentry/icons';
+import {IconCursorArrow} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import EventView from 'sentry/utils/discover/eventView';
@@ -39,22 +38,6 @@ function ReplayMetaData({replayErrors, replayRecord, showDeadRageClicks = true}:
 
   return (
     <KeyMetrics>
-      <KeyMetricLabel>{t('Start Time')}</KeyMetricLabel>
-      <KeyMetricData>
-        {replayRecord ? (
-          <TimeContainer>
-            <IconCalendar color="gray300" size="sm" />
-            <TimeSince
-              date={replayRecord.started_at}
-              isTooltipHoverable
-              unitStyle="regular"
-            />
-          </TimeContainer>
-        ) : (
-          <HeaderPlaceholder width="80px" height="16px" />
-        )}
-      </KeyMetricData>
-
       {showDeadRageClicks && (
         <Fragment>
           <KeyMetricLabel>{t('Dead Clicks')}</KeyMetricLabel>
@@ -140,12 +123,6 @@ const ClickCount = styled(Count)`
   color: ${p => p.theme.gray300};
   display: flex;
   gap: ${space(0.75)};
-  align-items: center;
-`;
-
-const TimeContainer = styled('div')`
-  display: flex;
-  gap: ${space(1)};
   align-items: center;
 `;
